breakup skip types into more granular bits for easier inclusion in prompts
