+++ packages/SkipTypes/src/index.ts
@@ -7,7 +7,9 @@ export * from './query-types';
 export * from './agent-types';
 export * from './artifact-types';
 export * from './auth-types';
-export * from './component-types';
+export * from './root-spec';
+export * from './child-spec';
+export * from './data-requirements';
 export * from './component-option';
 export * from './runtime-types';
 export * from './shared';
